requirements.txt
flask
flask-mysql-connector
gunicorn
python-dotenv
psycopg2-binary